requirements.txt
Flask==0.12.2
pyjwt==1.5.3
bcrypt==3.1.4
sqlalchemy==1.1.14
Flask-Admin
Flask-WTF==0.14.2
